Add sucessful pairing flow, which will list the apps at the end.
This still needs a lot of work in order to be robust code, but for now it will sucessfully pair and list the current applications on the server. Some of the code looks a little ugly, and a few places where we validate the server certificates are not in place. Security is of little concern here at the moment however, we want a working client before refactoring.